refactor: 패키지 구조 MVC로 개편
