+++ src/main/java/lotto/model/WinningTicket.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model;
 
 public class WinningTicket {
