+++ src/main/java/lotto/model/utils/NumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model.utils;
 
 import java.util.List;
