+++ src/test/java/lotto/model/LottoNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model;
 
 import static org.assertj.core.api.Assertions.*;
