+++ src/test/java/lotto/model/LottoTicketsTest.java
@@ -1,5 +1,6 @@
-package lotto.domain;
+package lotto.model;
 
+import lotto.model.utils.RandomNumberGenerator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
