+++ src/test/java/lotto/model/MoneyTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
